refactor: 스케줄러로 변형, 검증 로직 추가

+++ helper-core/src/main/java/org/example/domain/exam/usecase/DeleteExamUseCase.java
@@ -1,7 +1,9 @@
 package org.example.domain.exam.usecase;
 
 import lombok.RequiredArgsConstructor;
+import org.example.domain.exam.service.CheckExamService;
 import org.example.domain.exam.service.CommandExamService;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -10,8 +12,12 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class DeleteExamUseCase {
     private final CommandExamService commandExamService;
+    private final CheckExamService checkExamService;
 
+    @Scheduled(cron = "0 0 0 * * *")
     public void execute() {
-        commandExamService.deleteAllExams();
+        if (checkExamService.checkExamWasFinished()) {
+            commandExamService.deleteAllExams();
+        }
     }
 }
